refactor: cache clear dialog
